// frontend/ClientApp/src/modules/contract-workflow-engine/models/contract-content-values-store.ts
import type { SemanticConditionValue } from '@/models/contract-data'

export interface ContractContentValuesState {
  semanticConditionValues: SemanticConditionValue[]
}

export type SemanticConditionValueSetter =
  | ((blockId: string, conditionId: string, parameterName: string, parameterValue: string | number, subBlockId?: string) => void)
  | null